[voicecall] Add multi sim enablers for VoiceCallManager (QML). Contributes to JB#33946
This commit adds an overloaded VoiceCallManager::dial method that does
not require provider argument rather the method discovers the
current needed provider based on current modemPath.

This allows us to have a clean binding on QML side for modemPath
that is based on current active SIM card.

This is not as clean as it would be. My personal take is that let's
cleanup this further if/when we drop telepathy ring from use.

The old VoiceCallManager::dial(const QString &provider, const QString &msisdn)
is kept for backwards compatibility reasons / for SIP calls.