VoiceCall: Make it work without Mer specific bits
Other distros such as LuneOS uses VoiceCall as well. To date LuneOS was
always maintaining patches to make sure that the Mer specific bits not
available in LuneOS were disabled.

Made the code more flexible so Voicecall can be used by simply enabling
or disabling some flags in the build and in this way a single codebase
can be shared between Mer and LuneOS for Voicecall.

Signed-off-by: Herman van Hazendonk github.com@herrie.org