[voicall] Add a filtering action for incoming calls.
When an incoming call arrives, besides existing
answer() and hangup() responses, this commit
adds a new one: filter(action). The action
filter() is doing, can be to block the call
(hangup automatically) or to ignore the call
(call is not hangup but no feedback is given).

This introduce two new states, REJECTED and
IGNORED.