[voicecall] Delay saving event for incoming calls.
The incoming status (received or missed) in
not yet known when a new call arrives. Delay
the event storage so the proper incoming
status is known when saving.

This is required for instance in libcommhistory
event models, since the missed call counter is
updated on event addition, not on event update.