[voicecall] Add setters to the filter plugin.
Expand the filter API to add semantic
functions to manipulate ignored and
rejected numbers or range of numbers.

This is packaged in a library to help
downstream packages to handle the filters.

It also adds a test, using this new API,
to ensure stability and conformance of
the filtering plugin.