[voicecall] Use a temp file for ongoing call recordings. JB#62549
The .wav file is valid only after writing the header at the end of the
recording. Also the current UI having FileModel doesn't cope well
with file sizes updating live.

While at it, adjusted the file name construction to use one pass
for placeholder replacements. In theory the external input might
include %N that would confuse the latter .arg()s. And even if not
happening for this easily, good to promote the single pass as practice.